requirements.txt
streamlit
openai
pydantic
python-dotenv
rank-bm25